+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `Date.parseDuration(str)` for getting a duration in ms
 * `Date#add` and `Date#subtract` now also accept duration strings
 * Total `RURL` rewrite, inspired by `url-parse` by Arnout Kazemier and `URI.js` by Rodney Rehm
+* `Object.flatten(obj, divider)` now accepts custom dividers
 
 ## 0.5.6 (2018-06-20)